move isTranslation and replaceFirstTranslation to filter.js
